refactor: pulls complexity down
