+++ src/core_modules/capture-core-utils/RulesEngine/RulesEngine.test.js
@@ -1,8 +1,5 @@
 // @flow
 import RulesEngine from '../RulesEngine/RulesEngine';
-import inputValueConverter from '../../capture-core/rulesEngineActionsCreator/converters/inputValueConverter';
-import momentConverter from '../../capture-core/rulesEngineActionsCreator/converters/momentConverter';
-import outputRulesEffectsValueConverter from '../../capture-core/rulesEngineActionsCreator/converters/rulesEffectsValueConverter';
 import runRulesForSingleEvent from '../../capture-core/rulesEngineActionsCreator/runRulesForSingleEvent';
 
 const programs = [
@@ -42,7 +39,7 @@ const programs = [
 
 programs.forEach(({ program, foundation, orgUnit }) => {
     test('Tests on runRulesForSingleEvent function', () => {
-        const rulesEngine = new RulesEngine(inputValueConverter, momentConverter, outputRulesEffectsValueConverter);
+        const rulesEngine = new RulesEngine();
 
         const rulesEffects = runRulesForSingleEvent(rulesEngine, program, foundation, orgUnit);
