+++ src/core_modules/capture-core-utils/RulesEngine/converters/momentConverter.js
@@ -1,7 +1,7 @@
 // @flow
 import moment from 'moment';
-import { getFormattedStringFromMomentUsingEuropeanGlyphs } from 'capture-core-utils/date';
-import type { IMomentConverter } from 'capture-core-utils/RulesEngine/rulesEngine.types';
+import { getFormattedStringFromMomentUsingEuropeanGlyphs } from '../../date';
+import type { IMomentConverter } from '../rulesEngine.types';
 
 const momentFormat = 'YYYY-MM-DD';
