+++ src/core_modules/capture-core-utils/RulesEngine/converters/rulesEffectsValueConverter.js
@@ -1,8 +1,8 @@
 // @flow
 /* eslint-disable class-methods-use-this */
 import moment from 'moment';
-import { convertMomentToDateFormatString } from '../../utils/converters/date';
-import type { IConvertOutputRulesEffectsValue } from '../RulesEngine/rulesEngine.types';
+import { convertMomentToDateFormatString } from '../../../capture-core/utils/converters/date';
+import type { IConvertOutputRulesEffectsValue } from '../rulesEngine.types';
 
 const dateMomentFormat = 'YYYY-MM-DD';
